feat(2.0): consolidate CLI to six unified commands
Replace the previous ten command files with six new commands that share
a uniform "drill in by adding flags" UX vocabulary:

- team: fetchTeams (no flags) / fetchSquad (-s + --name) / fetchLineup
  (-s -r). One verb covers what was three commands.
- player: fetchPlayerDetails. Renamed from player-details.
- match: fetchMatches with --status filter. Subsumes matches + fixture.
- stats: --by player (default) → fetchPlayerStats; --by team →
  fetchTeamStats. Subsumes team-stats.
- ladder: rewritten using the command-builder; surface unchanged.
- awards: --type {brownlow,coleman,coaches,all-australian,rising-star}.
  Hides fetched vs computed source heterogeneity. Subsumes coaches-votes.

src/cli.ts updated to register the six new subCommands and remove the
ten old registrations.

Deleted command files:
  matches.ts, fixture.ts, teams.ts, squad.ts, lineup.ts, team-stats.ts,
  coaches-votes.ts, player-details.ts.

Quality gates: typecheck, lint, 271 tests pass.
