+++ src/cli/commands/ladder.ts
@@ -1,16 +1,12 @@
-import { defineCommand } from "citty";
+/**
+ * `ladder` command — season standings.
+ */
+
 import { fetchLadder } from "../../index";
-import { withErrorBoundary } from "../error-boundary";
+import { defineFitzroyCommand } from "../command-builder";
 import { COMPETITION_FLAG, OUTPUT_FLAGS, ROUND_FLAG, SEASON_FLAG, SOURCE_FLAG } from "../flags";
-import { type FormatOptions, formatOutput, type TableColumnConfig } from "../formatters/index";
-import { showSummary, withSpinner } from "../ui";
-import {
-  validateCompetition,
-  validateFormat,
-  validateRound,
-  validateSeason,
-  validateSource,
-} from "../validation";
+import type { TableColumnConfig } from "../formatters/index";
+import { validateCompetition, validateRound, validateSeason, validateSource } from "../validation";
 
 const DEFAULT_COLUMNS: TableColumnConfig[] = [
   { key: "position", label: "Pos", maxWidth: 4 },
@@ -22,7 +18,14 @@ const DEFAULT_COLUMNS: TableColumnConfig[] = [
   { key: "premiershipsPoints", label: "Pts", maxWidth: 5 },
 ];
 
-export const ladderCommand = defineCommand({
+interface LadderArgs {
+  season: string;
+  round?: string;
+  source: string;
+  competition: string;
+}
+
+export const ladderCommand = defineFitzroyCommand<LadderArgs & Record<string, unknown>, object>({
   meta: {
     name: "ladder",
     description: "Fetch ladder standings for a season",
@@ -34,34 +37,19 @@ export const ladderCommand = defineCommand({
     ...COMPETITION_FLAG,
     ...OUTPUT_FLAGS,
   },
-  run: withErrorBoundary(async ({ args }) => {
+  columns: DEFAULT_COLUMNS,
+  spinner: "Fetching ladder…",
+  run: async (args) => {
     const season = validateSeason(args.season);
     const round = args.round ? validateRound(args.round) : undefined;
     const source = validateSource(args.source);
     const competition = validateCompetition(args.competition);
-    const format = validateFormat(args.format);
 
-    const result = await withSpinner("Fetching ladder…", () =>
-      fetchLadder({ source, season, round, competition }),
-    );
-
-    if (!result.success) {
-      throw result.error;
-    }
-
-    const data = result.data;
-    showSummary(
-      `Loaded ladder for ${season}${round ? ` round ${round}` : ""} (${data.entries.length} teams)`,
-    );
-
-    const formatOptions: FormatOptions = {
-      json: args.json,
-      csv: args.csv,
-      format,
-      full: args.full,
-      columns: DEFAULT_COLUMNS,
-    };
-
-    console.log(formatOutput(data.entries, formatOptions));
-  }),
+    const result = await fetchLadder({ source, season, round, competition });
+    if (!result.success) return result;
+    // Surface the entries[] for tabular output.
+    return { success: true, data: [...result.data.entries] };
+  },
+  summary: (data, args) =>
+    `Loaded ladder for ${args.season}${args.round ? ` round ${args.round}` : ""} (${data.length} teams)`,
 });
